Feature/apple review: Make app ready for review from apple (#22)
* feat: add dartdoc and index generator configuration files

* add initial implementation of report generation and history analysis

* feat: add apple sign in

* feat: account deletion

* feat: add open source licenses

* feat: add consent and privacy policy

* feat: add untis demo mode

+++ lib/routes/navigation_shell.dart
@@ -20,7 +20,7 @@ class NavigationShell extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    int maxWidth = context.select(
+    double maxWidth = context.select(
       (ConfigProvider provider) => provider.maxWidthThreshold,
     ); // Define the maximum width for the layout
